discovery/kubernetes: extract service endpoint discovery
This extract discovery of services and their endpoints into its own
type.
